requirements.txt
Django==4.2.4
django_environ==0.10.0
djangorestframework==3.14.0
python-decouple==3.8
whitenoise==6.5.0
gunicorn==21.2.0
pandas==2.0.3
